+++ src/test/java/org/jenkinsci/plugins/jvcts/perform/JvctsPerformerParseTest.java
@@ -2,6 +2,7 @@ package org.jenkinsci.plugins.jvcts.perform;
 
 import static com.google.common.base.Joiner.on;
 import static java.nio.charset.Charset.defaultCharset;
+import static org.jenkinsci.plugins.jvcts.perform.JvctsPerformer.doPerform;
 import static org.jenkinsci.plugins.jvcts.stash.StashClientFaker.COMMENTS_1_DELETE;
 import static org.jenkinsci.plugins.jvcts.stash.StashClientFaker.COMMENTS_2_DELETE;
 import static org.jenkinsci.plugins.jvcts.stash.StashClientFaker.COMMENTS_CHECKSTYLEFILE_GET;
@@ -10,9 +11,9 @@ import static org.jenkinsci.plugins.jvcts.stash.StashClientFaker.COMMENTS_PMDAND
 import static org.jenkinsci.plugins.jvcts.stash.StashClientFaker.COMMENTS_PMDFILE_GET;
 import static org.jenkinsci.plugins.jvcts.stash.StashClientFaker.fakeStashClient;
 import static org.jenkinsci.plugins.jvcts.stash.StashClientFaker.getRequestsSentToStash;
+import static org.jenkinsci.plugins.jvcts.stash.StashClientFaker.prToCommit;
 import static org.jenkinsci.plugins.jvcts.stash.StashClientFaker.readFile;
 import static org.junit.Assert.fail;
-import hudson.model.BuildListener;
 import hudson.model.StreamBuildListener;
 
 import java.io.File;
@@ -23,7 +24,6 @@ import java.util.logging.Logger;
 import org.jenkinsci.plugins.jvcts.config.Parser;
 import org.jenkinsci.plugins.jvcts.config.ParserConfig;
 import org.jenkinsci.plugins.jvcts.config.ViolationsToStashConfig;
-import org.junit.Before;
 import org.junit.Test;
 
 import com.google.common.io.Resources;
@@ -31,8 +31,7 @@ import com.google.common.io.Resources;
 public class JvctsPerformerParseTest {
  private ViolationsToStashConfig config;
 
- @Before
- public void before() throws IOException {
+ public void preConfigure() throws IOException {
   config = new ViolationsToStashConfig();
   ParserConfig checkStyleParserConfig = new ParserConfig(Parser.CHECKSTYLE.getTypeDescriptorName(), "**/"
     + Parser.CHECKSTYLE.getTypeDescriptorName() + ".xml, **/" + Parser.CHECKSTYLE.getTypeDescriptorName()
@@ -48,15 +47,12 @@ public class JvctsPerformerParseTest {
   config.setStashUser("stashUser");
   config.setStashPassword("stashPassword");
   config.setStashProject("stashProject");
-  config.setStashPullRequestId("1");
   config.setStashRepo("stashRepo");
 
   disableLogging();
 
   fakeStashClient();
 
-  BuildListener listener = new StreamBuildListener(System.out, defaultCharset());
-  JvctsPerformer.doPerform(config, getWorkspace(), listener);
  }
 
  private void disableLogging() {
@@ -67,14 +63,20 @@ public class JvctsPerformerParseTest {
  }
 
  @Test
- public void testThatCheckstyleIsCommented() throws IOException {
+ public void testThatPullRequestCheckstyleIsCommented() throws IOException {
+  preConfigure();
+  config.setStashPullRequestId("1");
+  doPerform(config, getWorkspace(), new StreamBuildListener(System.out, defaultCharset()));
   assertRequested(readFile("checkstyle_checkstylefile_1.json"));
   assertRequested(readFile("checkstyle_checkstylefile_2.json"));
   assertRequested(readFile("checkstyle_checkstylefile_3_relativePath.json"));
  }
 
  @Test
- public void testThatOldCommentsAreDeleted() throws IOException {
+ public void testThatPullRequestOldCommentsAreDeleted() throws IOException {
+  preConfigure();
+  config.setStashPullRequestId("1");
+  doPerform(config, getWorkspace(), new StreamBuildListener(System.out, defaultCharset()));
   assertRequested(COMMENTS_CHECKSTYLEFILE_GET);
   assertRequested(COMMENTS_PMDANDCHECKSTYLE_GET);
   assertRequested(COMMENTS_PMDFILE_GET);
@@ -84,18 +86,74 @@ public class JvctsPerformerParseTest {
  }
 
  @Test
- public void testThatCheckstyleAndPmdCanCommentsOnSameFileIsCommented() throws IOException {
+ public void testThatPullRequestCheckstyleAndPmdCanCommentsOnSameFileIsCommented() throws IOException {
+  preConfigure();
+  config.setStashPullRequestId("1");
+  doPerform(config, getWorkspace(), new StreamBuildListener(System.out, defaultCharset()));
   assertRequested(readFile("checkstyle_pmdandcheckstyle.json"));
  }
 
  @Test
- public void testThatPmdIsCommented() throws IOException {
+ public void testThatPullRequestPmdIsCommented() throws IOException {
+  preConfigure();
+  config.setStashPullRequestId("1");
+  doPerform(config, getWorkspace(), new StreamBuildListener(System.out, defaultCharset()));
   assertRequested(readFile("pmd_pmdfile.json"));
  }
 
  @Test
- public void testThatFindbugsIsCommented() throws IOException {
+ public void testThatPullRequestFindbugsIsCommented() throws IOException {
+  preConfigure();
+  config.setStashPullRequestId("1");
+  doPerform(config, getWorkspace(), new StreamBuildListener(System.out, defaultCharset()));
   assertRequested(readFile("findbugs_code.json"));
+ }
+
+ @Test
+ public void testThatCommitCheckstyleIsCommented() throws IOException {
+  preConfigure();
+  config.setCommitHash("1");
+  doPerform(config, getWorkspace(), new StreamBuildListener(System.out, defaultCharset()));
+  assertRequested(prToCommit(readFile("checkstyle_checkstylefile_1.json")));
+  assertRequested(prToCommit(readFile("checkstyle_checkstylefile_2.json")));
+  assertRequested(prToCommit(readFile("checkstyle_checkstylefile_3_relativePath.json")));
+ }
+
+ @Test
+ public void testThatCommitOldCommentsAreDeleted() throws IOException {
+  preConfigure();
+  config.setCommitHash("1");
+  doPerform(config, getWorkspace(), new StreamBuildListener(System.out, defaultCharset()));
+  assertRequested(prToCommit(COMMENTS_CHECKSTYLEFILE_GET));
+  assertRequested(prToCommit(COMMENTS_PMDANDCHECKSTYLE_GET));
+  assertRequested(prToCommit(COMMENTS_PMDFILE_GET));
+  assertRequested(prToCommit(COMMENTS_FINDBUGS_GET));
+  assertRequested(prToCommit(COMMENTS_1_DELETE));
+  assertRequested(prToCommit(COMMENTS_2_DELETE));
+ }
+
+ @Test
+ public void testThatCommitCheckstyleAndPmdCanCommentsOnSameFileIsCommented() throws IOException {
+  preConfigure();
+  config.setCommitHash("1");
+  doPerform(config, getWorkspace(), new StreamBuildListener(System.out, defaultCharset()));
+  assertRequested(prToCommit(readFile("checkstyle_pmdandcheckstyle.json")));
+ }
+
+ @Test
+ public void testThatCommitPmdIsCommented() throws IOException {
+  preConfigure();
+  config.setCommitHash("1");
+  doPerform(config, getWorkspace(), new StreamBuildListener(System.out, defaultCharset()));
+  assertRequested(prToCommit(readFile("pmd_pmdfile.json")));
+ }
+
+ @Test
+ public void testThatCommitFindbugsIsCommented() throws IOException {
+  preConfigure();
+  config.setCommitHash("1");
+  doPerform(config, getWorkspace(), new StreamBuildListener(System.out, defaultCharset()));
+  assertRequested(prToCommit(readFile("findbugs_code.json")));
  }
 
  private File getWorkspace() {